+++ changelog.md
@@ -9,6 +9,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Changed
+
+- No more checking if lambda.bx exists on each execution, which accelerates the execution of the lambda runtime.
+
+### Added
+
+- Listener support so you can add an 'Application.bx/cfc' to your lambda runtime and listen to events from the lambda runtime life-cycle.
+- Multi-Function execution support via the `x-bx-function` request header.  This will allow you to build RESTFul APIs with multiple functions in a single Lambda runtime.
+- Modules support. Allow to load modules into Lambda runtimes via the `modules` convention
+- `box.json` support for Lambda runtimes by convention in the lambda root
+- More testing facilities for Lambda runtimes
+
 ## [1.0.0-beta26] - 2025-01-18
 
 ## [1.0.0-beta25] - 2024-12-13